Bullets now spawn the correct place, and move according the speed.
Bullet spawns from the center of the TankBase. Then continues from this
point

Created class Point for double x and y.

One problem so far is the tank can turn to make the shot look stupid (it
continues from a point not intended for the bullet)
To reach a satisfying result, we could make the tank stop for a moment,
until the bullet is fired. This would also mean there's a consequence
for shooting.

Throttling for the bullets is also missing