TL;DR The camera now works as an active part of the game.
By using offsets, the map is now rendered according to the position of
the player. By this I mean, with the player is in the center of the map,
and the offset would be player x - mapwidth/2. The tiles then gets this
offset added to them, to afterwards be rendered according to its new
position.

Other than that, the constraints have been fixed with this offset logic.

Tile rendering was fixed. This was caused by the if statement not
checking the right variables when choosing which tiles to render.