Fixed rendering issues.
It would render the right side correctly, but not calculate the start point of the drawing. So if the were at the bottom right corner of the map (since the rendering starts from the top-left corner and goes right) the GPU would struggle to keep up. This issue has been fixed